Fix random test failures on Mac (#4)
Using `tempy` to generate random temporary directories results in
unix domain socket names that are too long (104 bytes is the max
on Mac) which manifests itself as EADDRINUSE errors. So replace
`tempy` with a simple alternative.

Also expands the CI test matrix, adding Node.js 18 and Windows and
Mac, to test platform differences in the handling of sockets.